requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
anthropic>=0.40.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
streamlit>=1.40.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.18.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
